requirements.txt
django-autoslug
pillow
